requirements.txt
openai
python-dotenv
requests
autogen
autogen-ext[openai]
openmeteo-requests
azure-cosmos
requests-cache
retry-requests
pytest
streamlit-cookies-manager
fastapi 
uvicorn[standard]
python-multipart
gunicorn
azure.cognitiveservices.speech
pydub
azure-storage-blob
langchain
 langchain-openai
langchain-community
langchain-azure-ai